sphinxsort decomposition part2
